requirements.txt
absl-py==0.11.0
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.3.1
astroid==2.4.2
astunparse==1.6.3
autopep8==1.5.5
beautifulsoup4==4.9.3
cachetools==4.2.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
defusedxml==0.6.0
distlib==0.3.1
Django==3.1.4
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-quill-editor==0.1.16
django-seed==0.2.2
django-taggit==1.3.0
djangorestframework==3.12.2
Faker==6.0.0
filelock==3.0.12
google-auth==1.27.1
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
html5lib==1.1
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.20.1
oauthlib==3.1.0
Pillow==8.1.0
promise==2.3
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pydub==0.25.1
python-crontab==2.5.1
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2020.5
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.7
Rx==1.6.1
schedule==1.0.0
singledispatch==3.4.0.3
six==1.15.0
slugify==0.0.1
snowballstemmer==2.0.0
soupsieve==2.0.1
SQLAlchemy==1.3.23
sqlparse==0.4.1
tensorboard-plugin-wit==1.8.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
